+++ src/app.js
@@ -1,173 +1,11 @@
 import './app.scss';
 import { createElement } from './lib/dom';
 import { createTitle } from './components/title';
-import { createSearch } from './components/search';
-import { createPokemons } from './components/pokemons';
+import { createSearchInput } from './components/search';
+import { createSearchResults } from './components/pokemons';
 import { appendContent } from './lib/dom';
 import Logo from './assets/fav_pichu.png';
-
-const allPokemons = [
-  'Bulbasaur',
-  'Ivysaur',
-  'Venusaur',
-  'Charmander',
-  'Charmeleon',
-  'Charizard',
-  'Squirtle',
-  'Wartortle',
-  'Blastoise',
-  'Caterpie',
-  'Metapod',
-  'Butterfree',
-  'Weedle',
-  'Kakuna',
-  'Beedrill',
-  'Pidgey',
-  'Pidgeotto',
-  'Pidgeot',
-  'Rattata',
-  'Raticate',
-  'Spearow',
-  'Fearow',
-  'Ekans',
-  'Arbok',
-  'Pikachu',
-  'Raichu',
-  'Sandshrew',
-  'Sandslash',
-  'Nidoran',
-  'Nidorina',
-  'Nidoqueen',
-  'Nidoran',
-  'Nidorino',
-  'Nidoking',
-  'Clefairy',
-  'Clefable',
-  'Vulpix',
-  'Ninetales',
-  'Jigglypuff',
-  'Wigglytuff',
-  'Zubat',
-  'Golbat',
-  'Oddish',
-  'Gloom',
-  'Vileplume',
-  'Paras',
-  'Parasect',
-  'Venonat',
-  'Venomoth',
-  'Diglett',
-  'Dugtrio',
-  'Meowth',
-  'Persian',
-  'Psyduck',
-  'Golduck',
-  'Mankey',
-  'Primeape',
-  'Growlithe',
-  'Arcanine',
-  'Poliwag',
-  'Poliwhirl',
-  'Poliwrath',
-  'Abra',
-  'Kadabra',
-  'Alakazam',
-  'Machop',
-  'Machoke',
-  'Machamp',
-  'Bellsprout',
-  'Weepinbell',
-  'Victreebel',
-  'Tentacool',
-  'Tentacruel',
-  'Geodude',
-  'Graveler',
-  'Golem',
-  'Ponyta',
-  'Rapidash',
-  'Slowpoke',
-  'Slowbro',
-  'Magnemite',
-  'Magneton',
-  "Farfetch'd",
-  'Doduo',
-  'Dodrio',
-  'Seel',
-  'Dewgong',
-  'Grimer',
-  'Muk',
-  'Shellder',
-  'Cloyster',
-  'Gastly',
-  'Haunter',
-  'Gengar',
-  'Onix',
-  'Drowzee',
-  'Hypno',
-  'Krabby',
-  'Kingler',
-  'Voltorb',
-  'Electrode',
-  'Exeggcute',
-  'Exeggutor',
-  'Cubone',
-  'Marowak',
-  'Hitmonlee',
-  'Hitmonchan',
-  'Lickitung',
-  'Koffing',
-  'Weezing',
-  'Rhyhorn',
-  'Rhydon',
-  'Chansey',
-  'Tangela',
-  'Kangaskhan',
-  'Horsea',
-  'Seadra',
-  'Goldeen',
-  'Seaking',
-  'Staryu',
-  'Starmie',
-  'Mr. Mime',
-  'Scyther',
-  'Jynx',
-  'Electabuzz',
-  'Magmar',
-  'Pinsir',
-  'Tauros',
-  'Magikarp',
-  'Gyarados',
-  'Lapras',
-  'Ditto',
-  'Eevee',
-  'Vaporeon',
-  'Jolteon',
-  'Flareon',
-  'Porygon',
-  'Omanyte',
-  'Omastar',
-  'Kabuto',
-  'Kabutops',
-  'Aerodactyl',
-  'Snorlax',
-  'Articuno',
-  'Zapdos',
-  'Moltres',
-  'Dratini',
-  'Dragonair',
-  'Dragonite',
-  'Mewtwo',
-  'Mew'
-];
-
-function filterPokemons(searchValue) {
-  const lowerCaseSearchValue = searchValue.toLowerCase();
-
-  const filteredPokemons = allPokemons.filter(pokemon => {
-    return pokemon.toLowerCase().includes(lowerCaseSearchValue);
-  });
-  return filteredPokemons;
-}
+import { filterPokemons } from './lib/pokemonsTwo';
 
 export function app() {
   const header = createElement('header', {
@@ -177,31 +15,33 @@ export function app() {
     className: 'main'
   });
   const title = createTitle('Pokédex');
-  const searchInput = createSearch(sessionStorage.getItem('searchValue'));
+  const searchInput = createSearchInput({
+    value: sessionStorage.getItem('searchValue')
+  });
   const logo = createElement('img', {
     className: 'logo',
     src: Logo
   });
 
-  let pokemons = null;
+  let searchResults = null;
   function setSearchResults() {
     const filteredPokemons = filterPokemons(searchInput.value);
-    pokemons = createPokemons(filteredPokemons);
-    appendContent(main, pokemons);
+    searchResults = createSearchResults({
+      items: filteredPokemons
+    });
+    appendContent(main, searchResults);
   }
   setSearchResults();
 
   appendContent(header, [title]);
-  appendContent(main, [logo, searchInput, pokemons]);
+  appendContent(main, [logo, searchInput, searchResults]);
 
   searchInput.addEventListener('input', event => {
-    main.removeChild(pokemons);
+    main.removeChild(searchResults);
     setSearchResults();
 
     const searchValue = event.target.value;
-
     sessionStorage.setItem('searchValue', searchValue);
   });
-
   return [header, main];
 }
